Added StellarisWare, RASLib, and RASDemo
RASLib is the 2010 version. RASDemo is a new project that copies all of
its code from the 2010 RobzDemo. All I did was change the name (it was
obviously a terrible name to begin with) and made it as a uvproj (versus
a uv2, and seeing as we're using uVision 4...). After RASDemo built and
ran on the board, I noticed that the encoders weren't working, but fixed
this by copying over RobzDemo's Startup.S (DUH).